+++ app/src/main/java/com/nexohogar/domain/model/AccountBalance.kt
@@ -1,11 +1,8 @@
 package com.nexohogar.domain.model
 
-/**
- * Modelo de dominio puro para el balance de una cuenta.
- */
 data class AccountBalance(
     val accountId: String,
     val accountName: String,
     val accountType: String,
-    val movementBalance: Double
+    val movementBalance: Long
 )
